+++ src/Topxia/WebBundle/Controller/CourseReviewController.php
@@ -59,6 +59,8 @@ class CourseReviewController extends CourseBaseController
     {
         $user = $this->getCurrentUser();
 
+        list($course, $member) = $this->getCourseService()->tryTakeCourse($id);
+
         $fields             = $request->request->all();
         $fields['userId']   = $user['id'];
         $fields['courseId'] = $id;
@@ -96,6 +98,16 @@ class CourseReviewController extends CourseBaseController
 
     public function deleteAction(Request $request, $reviewId)
     {
+        $user = $this->getCurrentUser();
+        if (!$user->isLogin()) {
+            throw $this->createAccessDeniedException('not login');
+        }
+
+        $review = $this->getReviewService()->getReview($reviewId);
+        if ($review['userId'] != $user['id']) {
+            throw $this->createAccessDeniedException('review is not exsits.');
+        }
+
         $this->getReviewService()->deleteReview($reviewId);
         return $this->createJsonResponse(true);
     }
